Add files to the arena - the 'blessed' bluetooth stack, and pieces of the Nordic SDK
This commit adds the 'blessed' bluetooth stack along with some
examples, originally taken from here:
https://github.com/pauloborges/blessed

The project seems to be inactive now. I decided to copy the repository,
instead of linking to it as a submodule.

To compile the library and the examples, some files from the
Nordic SDK are needed. I extracted some files of the Nordic SDK,
version 4.4.2, which is old, but it is nice that it is much smaller
than the latest SDKs. I only copied some header and source files,
I put them in the 'arena/nrf51822' directory.